Passing the first test (cheating)
Using the "POST" method against the "/api/instructions" url to create a new
instruction. And making sure all values sent were received and returne by the
server.

In this commit, the BuildInstruction.toBackbone() method's code is being almost
completely ignored to make this first test pass. I'll fix it in the next commits with
more tests.