Listing instructions
Just a small refactory in the controller that returns a list of instructions. I'm also
adding a test that ensures that one building instruction is being created.

The problem is that these tests are not well written, cause they currently depend
on the database instead of a mock. The first problem that happens because of
it is that I have to erase the db everytime before running the tests (yuck).