Fixing a synchronicity problem in the build runner
There were repositories being built and others were not. There was not a clear
difference between them, except it's time to build. So, after some hours, I
realized that there was a race condition in the access to redis to save the
"started date" field and the subscription to the event handling of the ChildProcess
instance.

To fix this, I just moved the spawn call to the callback called *after* saving
this data to the storage.