Calling the BuildRunner through the API instead of spawning a new command
This commit introduces a very nasty bug. Firstly, we need to kill programs that
we spawn, like git (when cloning) and the build script if they get out of control
(which is quite usual :P). To kill them without killing node interpreter, we need
to start a new session using setsid(), which is not supported by the spawn
command.

I'm currently using the `posix.setsid()` call, which poisons the process its first
call. This way, after this commit, I'll not be able to build a instruction more than
once!