A small hammer to propagate the custom settings path (Closes: #27)
Just fixing a hammer with another one. Emerald forks itself to run buils, but
without taking care about the customized settings file.

This patch fixes this problem by saving the path of the custom settings file and
passing to a centralized dispatcher that will be responsible for spawning all
emerald processes.